feat: add types in loading indicator page
